fix: make header more responsive (mobile)
